Remove conversion warning from include files 	modified:   include/geos/geom/BinaryOp.h 	modified:   include/geos/noding/BasicSegmentString.h

+++ include/geos/noding/BasicSegmentString.h
@@ -70,7 +70,7 @@ public:
 	/// see dox in SegmentString.h
 	unsigned int size() const override
 	{
-		return pts->size();
+		return static_cast<unsigned int>(pts->size());
 	}
 
 	/// see dox in SegmentString.h
